Add the `add-all` command
Nobody has time to find all their git repos first! The `add-all` command
finds all git repos within the current directory and adds them to the
list of tracked repos.

For example, given the following directory structure, running
`systemgit add-all` from the `code` directory would add all of the
projects.

    code/
        project1/
        go/
            project2/
            project3/
        js/
            project4/

Note that we have to be careful not to assume that the user has called
the script `systemgit`, so `$0` is used to run `systemgit add` from each
found directory.